requirements.txt
google-generativeai>=0.3.0
python-dotenv>=1.0.0
numpy>=1.20.0
pillow>=9.0.0
pybullet>=3.2.0
matplotlib>=3.5.0
pyyaml>=6.0
sentence-transformers>=2.0.0
